requirements.txt
streamlit>=1.28.0
networkx>=3.1
requests>=2.28.0
google-generativeai>=0.3.0
matplotlib>=3.6.0
python-dotenv>=1.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
scikit-learn>=1.3.0
numpy>=1.24.0